+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to the "Diff Folders" extension will be documented in this f
 - Fixed missing description in README for `Go to File`.
 - Fixed open to side for multiple files.
 
+### Changed
+- Disabled ui if files will be deleted.
+- Disabled drag and drop for error und unknown files.
+
 ## [0.30.0] - 2020-11-01
 
 ### Added